feat: add booking page price display and GET /api/pitches/{id}
- Add GET /api/pitches/{id} (PitchesController, IPitchService,
  PitchService, PitchResponse) so the booking page can fetch pitch
  metadata when opened by direct URL.
- BookingPage: fetch pitch on mount, default currency to $, show price
  under each duration chip, update footer total as soon as duration
  changes.
- Fix /api/bookings/my 500 — map Postgres COUNT(*) OVER() as long in
  BookingWithCount (was int, broke Dapper record materialization once
  a player had any bookings).

+++ backend/SmartSports.DAL/Repositories/BookingRepository.cs
@@ -194,7 +194,7 @@ public class BookingRepository : IBookingRepository
             PitchName = r.PitchName
         });
 
-        var totalCount = list.FirstOrDefault()?.TotalCount ?? 0;
+        var totalCount = (int)(list.FirstOrDefault()?.TotalCount ?? 0);
 
         return (items, totalCount);
     }
@@ -272,7 +272,7 @@ public class BookingRepository : IBookingRepository
             PitchName = r.PitchName
         });
 
-        var totalCount = list.FirstOrDefault()?.TotalCount ?? 0;
+        var totalCount = (int)(list.FirstOrDefault()?.TotalCount ?? 0);
 
         return (items, totalCount);
     }
@@ -306,5 +306,5 @@ public class BookingRepository : IBookingRepository
         int Id, int UserId, int PitchId,
         DateOnly BookingDate, TimeOnly StartTime, TimeOnly EndTime,
         decimal TotalPrice, string Status, DateTime BookedAt,
-        string PitchName, int TotalCount);
+        string PitchName, long TotalCount);
 }
